requirements.txt
Mastodon.py==1.3.1
python-dotenv==0.9.1

